v0.17.149：修复 148 编译错（stats_page 进货视图'天数' _days.length int 赋给 double → toDouble）+ .gitignore 忽略 _tmp/（本地调试脚本统一存放不上传）

+++ frontend-flutter/lib/pages/stats_page.dart
@@ -665,7 +665,7 @@ class _StatsPageState extends State<StatsPage> {
     final data = _isBuy
         ? <String, double>{
             '进货': sales,
-            if (_mode != 'year') '天数': _days.length,
+            if (_mode != 'year') '天数': _days.length.toDouble(),
           }
         : <String, double>{
             '出货': sales,
